requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-chroma
langchain-ollama
langchain-huggingface
langchain-text-splitters
agno
sentence-transformers